everything: change intention of this repo
Previously, this was the place of development for an advancing operating
system. This system has been moved to https://codeberg.org/loup-os.
However, I want to help starters to learn how operating systems work,
and this is the new intention of this repository.

I understand that this is a big change.